requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
python-multipart
PyJWT[crypto]
pydantic
pytest
httpx
jinja2
